+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - fixed `idc.imgaug.filter.transfer_region` method: image segmentation layers now use values (0,255)
+- using underscores now instead of dashes in dependencies (`setup.py`)
 
 
 0.0.8 (2025-03-14)