+++ CHANGES.rst
@@ -13,7 +13,7 @@ Changelog
 0.0.8 (2025-03-14)
 ------------------
 
-- added support for placeholers: `idc-combine-sub-images`
+- added support for placeholders: `idc-combine-sub-images`
 
 
 0.0.7 (2025-02-26)